requirements.txt
asgiref==3.3.4
attrs==21.1.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dateparser==1.0.0
Django==3.2.2
django-coinpayments==0.1.3
django-crispy-forms==1.11.2
django-mathfilters==1.0.0
django-model-utils==4.1.1
humanize==3.5.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
maya==0.6.1
pendulum==2.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
regex==2021.4.4
service-identity==18.1.0
six==1.16.0
snaptime==0.2.4
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
zope.interface==5.4.0
